Store service settings in database
Added new entity Setting and a wrapper class ServiceSettingsProvider to wrap the relevant service settings. This class is used by Admin and Service project to read service settings from the database.

Introduced a new generic class DatabaseDatasource that is used when accessing the settings table.

Cleaned up application settings of service and admin.